Updated "registration_form.php" to include a "$root" variable that includes a path to the root directory for the project and updated "navbar_logged_out.php" to reference the new "$root" variable.
Updated "registration_form.php" to include a "$root" variable that includes a path to the root directory for the project and updated "navbar_logged_out.php" to reference the new "$root" variable. This will allow for included headers to work without having to worry about breakign relative links due to the headers being included in directories that may be nested, etc.